:bug: Respond to LP welcome in Composer template
Previously the Composer template was showing the welcome selection if a new conversation is created. Direct Line Connector does not support this event and won't show these messages. So instead the initial messages will not be shown when the bot receives the LP native welcome action inside channelData (on chat/transfer)